# docs/api/control-api.openapi.yaml
openapi: 3.0.3
info:
  title: DAEMON CHOIR Control API
  version: 2.0.0
  description: |
    Production REST API for controlling, configuring, and inspecting the local DAEMON CHOIR daemon.
    All endpoints bind to localhost:9876 loopback ONLY.
servers:
  - url: http://127.0.0.1:9876/v1
    description: Local loopback only
paths:
...
  /config/reload:
    post:
      summary: Trigger hot reload of config file from disk
      responses:
        "202":
          description: Reload triggered successfully
        "409":
          description: Conflict, config contains errors
  /config/backends:
    put:
      summary: Dynamically update OSC backends
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  address:
                    type: string
      responses:
        "200":
          description: Backends updated
...
  /shutdown:
    post:
      summary: Graceful shutdown of the daemon
      parameters:
        - name: X-Daemon-Choir
          in: header
          required: true
          schema:
            type: string
            enum: [shutdown]
      responses:
        "202":
          description: Shutdown initiated
        "403":
          description: Forbidden (invalid or missing CSRF guard header)
...